Fix materialization with default encoding
